@@ -6,6 +6,8 @@ namespace Yiisoft\ErrorHandler\Tests\Renderer;
 
 use Exception;
 use HttpSoft\Message\Uri;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 use ReflectionClass;
@@ -130,9 +132,16 @@ final class HtmlRendererTest extends TestCase
         );
     }
 
+    #[WithoutErrorHandler]
     public function testRenderCallStackItemIfFileIsNotExistAndLineMoreZero(): void
     {
-        $this->assertEmpty($this->invokeMethod(new HtmlRenderer(), 'renderCallStackItem', [
+        $errorMessage = null;
+        set_error_handler(
+            static function (int $code, string $message) use (&$errorMessage) {
+                $errorMessage = $message;
+            }
+        );
+        $result = $this->invokeMethod(new HtmlRenderer(), 'renderCallStackItem', [
             'file' => 'not-exist',
             'line' => 1,
             'class' => null,
@@ -141,7 +150,11 @@ final class HtmlRendererTest extends TestCase
             'index' => 1,
             'isVendorFile' => false,
             'reflectionParameters' => [],
-        ]));
+        ]);
+        restore_error_handler();
+
+        $this->assertSame('', $result);
+        $this->assertSame('file(not-exist): Failed to open stream: No such file or directory', $errorMessage);
     }
 
     public function testRenderRequest(): void
@@ -171,7 +184,7 @@ final class HtmlRendererTest extends TestCase
         $this->assertSame('Error (' . ErrorException::class . ')', $name);
     }
 
-    public function createServerInformationLinkDataProvider(): array
+    public static function createServerInformationLinkDataProvider(): array
     {
         return [
             'not-exist' => [null, ''],
@@ -185,9 +198,7 @@ final class HtmlRendererTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider createServerInformationLinkDataProvider
-     */
+    #[DataProvider('createServerInformationLinkDataProvider')]
     public function testCreateServerInformationLink(?string $serverSoftware, string $expected): void
     {
         $renderer = new HtmlRenderer();
@@ -199,7 +210,7 @@ final class HtmlRendererTest extends TestCase
         $this->assertStringContainsString($expected, $renderer->createServerInformationLink($serverRequestMock));
     }
 
-    public function argumentsToStringValueDataProvider(): array
+    public static function argumentsToStringValueDataProvider(): array
     {
         return [
             'int' => [[1], '1'],
@@ -222,11 +233,7 @@ final class HtmlRendererTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider argumentsToStringValueDataProvider
-     *
-     * @param mixed $args
-     */
+    #[DataProvider('argumentsToStringValueDataProvider')]
     public function testArgumentsToString(array $args, string $expected): void
     {
         $renderer = new HtmlRenderer();
@@ -270,7 +277,7 @@ final class HtmlRendererTest extends TestCase
         ]));
     }
 
-    public function isVendorFileReturnFalseDataProvider(): array
+    public static function isVendorFileReturnFalseDataProvider(): array
     {
         return [
             'null' => [null],
@@ -279,9 +286,7 @@ final class HtmlRendererTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider isVendorFileReturnFalseDataProvider
-     */
+    #[DataProvider('isVendorFileReturnFalseDataProvider')]
     public function testIsVendorFileReturnFalse(?string $file): void
     {
         $this->assertFalse($this->invokeMethod(new HtmlRenderer(), 'isVendorFile', ['file' => $file]));
