@@ -34,6 +34,8 @@ use function trim;
 /**
  * `ThrowableResponseFactory` renders `Throwable` object
  * and produces a response according to the content type provided by the client.
+ *
+ * @deprecated Use {@see \Yiisoft\ErrorHandler\ThrowableResponseFactory} instead.
  */
 final class ThrowableResponseFactory implements ThrowableResponseFactoryInterface
 {
@@ -53,10 +55,10 @@ final class ThrowableResponseFactory implements ThrowableResponseFactoryInterfac
     private ?string $contentType = null;
 
     public function __construct(
-        private ResponseFactoryInterface $responseFactory,
-        private ErrorHandler $errorHandler,
-        private ContainerInterface $container,
-        HeadersProvider $headersProvider = null,
+        private readonly ResponseFactoryInterface $responseFactory,
+        private readonly ErrorHandler $errorHandler,
+        private readonly ContainerInterface $container,
+        ?HeadersProvider $headersProvider = null,
     ) {
         $this->headersProvider = $headersProvider ?? new HeadersProvider();
     }
