@@ -8,25 +8,54 @@ use Psr\Http\Message\ServerRequestInterface;
 use Throwable;
 use Yiisoft\ErrorHandler\ErrorData;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
+use Yiisoft\Http\Header;
 
 /**
  * Formats throwable into HTTP headers.
  */
 final class HeaderRenderer implements ThrowableRendererInterface
 {
-    public function render(Throwable $t, ServerRequestInterface $request = null): ErrorData
-    {
-        return new ErrorData('', ['X-Error-Message' => self::DEFAULT_ERROR_MESSAGE]);
+    /**
+     * @param string|null $contentType The content type to be set in the response header.
+     */
+    public function __construct(
+        private readonly ?string $contentType = null,
+    ) {
     }
 
-    public function renderVerbose(Throwable $t, ServerRequestInterface $request = null): ErrorData
+    public function render(Throwable $t, ?ServerRequestInterface $request = null): ErrorData
     {
-        return new ErrorData('', [
-            'X-Error-Type' => $t::class,
-            'X-Error-Message' => $t->getMessage(),
-            'X-Error-Code' => (string) $t->getCode(),
-            'X-Error-File' => $t->getFile(),
-            'X-Error-Line' => (string) $t->getLine(),
-        ]);
+        return new ErrorData(
+            '',
+            $this->addContentTypeHeader([
+                'X-Error-Message' => self::DEFAULT_ERROR_MESSAGE,
+            ]),
+        );
+    }
+
+    public function renderVerbose(Throwable $t, ?ServerRequestInterface $request = null): ErrorData
+    {
+        return new ErrorData(
+            '',
+            $this->addContentTypeHeader([
+                'X-Error-Type' => $t::class,
+                'X-Error-Message' => $t->getMessage(),
+                'X-Error-Code' => (string) $t->getCode(),
+                'X-Error-File' => $t->getFile(),
+                'X-Error-Line' => (string) $t->getLine(),
+            ]),
+        );
+    }
+
+    /**
+     * @param array<string, string|string[]> $headers
+     * @return array<string, string|string[]>
+     */
+    private function addContentTypeHeader(array $headers): array
+    {
+        if ($this->contentType !== null) {
+            $headers[Header::CONTENT_TYPE] = $this->contentType;
+        }
+        return $headers;
     }
 }
