@@ -8,6 +8,7 @@ use Psr\Http\Message\ServerRequestInterface;
 use Throwable;
 use Yiisoft\ErrorHandler\ErrorData;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
+use Yiisoft\Http\Header;
 
 use function json_encode;
 
@@ -16,7 +17,9 @@ use function json_encode;
  */
 final class JsonRenderer implements ThrowableRendererInterface
 {
-    public function render(Throwable $t, ServerRequestInterface $request = null): ErrorData
+    private const CONTENT_TYPE = 'application/json';
+
+    public function render(Throwable $t, ?ServerRequestInterface $request = null): ErrorData
     {
         return new ErrorData(
             json_encode(
@@ -24,11 +27,12 @@ final class JsonRenderer implements ThrowableRendererInterface
                     'message' => self::DEFAULT_ERROR_MESSAGE,
                 ],
                 JSON_THROW_ON_ERROR | JSON_UNESCAPED_SLASHES
-            )
+            ),
+            [Header::CONTENT_TYPE => self::CONTENT_TYPE],
         );
     }
 
-    public function renderVerbose(Throwable $t, ServerRequestInterface $request = null): ErrorData
+    public function renderVerbose(Throwable $t, ?ServerRequestInterface $request = null): ErrorData
     {
         return new ErrorData(
             json_encode(
@@ -40,8 +44,9 @@ final class JsonRenderer implements ThrowableRendererInterface
                     'line' => $t->getLine(),
                     'trace' => $t->getTrace(),
                 ],
-                JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_INVALID_UTF8_SUBSTITUTE | JSON_PARTIAL_OUTPUT_ON_ERROR
-            )
+                JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_INVALID_UTF8_SUBSTITUTE | JSON_PARTIAL_OUTPUT_ON_ERROR
+            ),
+            [Header::CONTENT_TYPE => self::CONTENT_TYPE],
         );
     }
 }
