@@ -1,13 +1,15 @@
 <?php
 
+use Psr\Http\Message\ServerRequestInterface;
 use Yiisoft\ErrorHandler\CompositeException;
 use Yiisoft\ErrorHandler\Exception\ErrorException;
+use Yiisoft\ErrorHandler\Renderer\HtmlRenderer;
 use Yiisoft\FriendlyException\FriendlyExceptionInterface;
 
 /**
- * @var $this \Yiisoft\ErrorHandler\Renderer\HtmlRenderer
- * @var $request \Psr\Http\Message\ServerRequestInterface|null
- * @var $throwable \Throwable
+ * @var $this HtmlRenderer
+ * @var $request ServerRequestInterface|null
+ * @var $throwable Throwable
  * @var $solutions string[]
  */
 
@@ -35,6 +37,16 @@ $exceptionMessage = $throwable->getMessage();
         <?= file_get_contents(__DIR__ . '/development.css') ?>
     </style>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700">
+    <noscript>
+        <style>
+            .call-stack-vendor-items, .functionArguments {
+                display: block !important;
+            }
+            .call-stack-vendor-collapse .flex-1 {
+                display: none;
+            }
+        </style>
+    </noscript>
 </head>
 <body<?= !empty($theme) ? " class=\"{$this->htmlEncode($theme)}\"" : '' ?>>
 <header>
@@ -196,6 +208,9 @@ $exceptionMessage = $throwable->getMessage();
     <?= file_get_contents(__DIR__ . '/highlight.min.js') ?>
 </script>
 <script>
+    const LIGHT_THEME = 'light-theme';
+    const DARK_THEME = 'dark-theme';
+
     window.onload = function() {
         const codeBlocks = document.querySelectorAll('.solution pre code,.codeBlock');
         const callStackItems = document.getElementsByClassName('call-stack-item');
@@ -353,15 +368,13 @@ $exceptionMessage = $throwable->getMessage();
         }
 
         function enableDarkTheme() {
-            document.body.classList.remove('light-theme');
-            document.body.classList.add('dark-theme');
-            setCookie('yii-exception-theme', 'dark-theme');
+            applyTheme(DARK_THEME);
+            setCookie('yii-exception-theme', DARK_THEME);
         }
 
         function enableLightTheme() {
-            document.body.classList.remove('dark-theme');
-            document.body.classList.add('light-theme');
-            setCookie('yii-exception-theme', 'light-theme');
+            applyTheme(LIGHT_THEME);
+            setCookie('yii-exception-theme', LIGHT_THEME);
         }
     };
 
@@ -370,28 +383,32 @@ $exceptionMessage = $throwable->getMessage();
     document.onmouseup = function() { document.getElementsByTagName('body')[0].classList.remove('mousedown'); }
 
     <?php if (empty($theme)): ?>
-    var theme = getCookie('yii-exception-theme');
-
+    const theme = getCookie('yii-exception-theme');
     if (theme) {
-        document.body.classList.add(theme);
+        applyTheme(theme);
     }
     <?php endif; ?>
 
+    function applyTheme(theme){
+        document.body.classList.remove(DARK_THEME, LIGHT_THEME);
+        document.body.classList.add(theme);
+    }
+
     function setCookie(name, value) {
-        var date = new Date(2100, 0, 1);
-        var expires = "; expires=" + date.toUTCString();
+        const date = new Date(2100, 0, 1);
+        const expires = "; expires=" + date.toUTCString();
 
         document.cookie = name + "=" + (value || "")  + expires + "; path=/";
     }
 
     function getCookie(name) {
-        var nameEQ = name + "=";
-        var ca = document.cookie.split(';');
+        const nameEQ = name + "=";
+        const ca = document.cookie.split(';');
 
-        for (var i=0; i < ca.length; i++) {
-            var c = ca[i];
-            while (c.charAt(0)==' ') c = c.substring(1,c.length);
-            if (c.indexOf(nameEQ) == 0) return c.substring(nameEQ.length,c.length);
+        for (let i=0; i < ca.length; i++) {
+            let c = ca[i];
+            while (c.charAt(0) === ' ') c = c.substring(1,c.length);
+            if (c.indexOf(nameEQ) === 0) return c.substring(nameEQ.length,c.length);
         }
 
         return null;
